Restructure project - Extract functions - Eliminate side effects
